Feat/dml delete and Remove PhysicalPlan (#56)
* feat: implement dml delete

* fix: data is inserted after deletion and the search is empty

* perf: kipdb version up

* ci

* refactor: Remove `PhysicalPlan` and convert `LogicPlan` to `Executor`

* feat: remove the coupling between DataBase and Storage

* docs: add Storage support statement and code usage instructions

* feat: Implement ddl `Drop`

* feat: Implement ddl `Truncate`

* feat: Implement dql `Count`

* feat: Implement dql `Sum`

* feat: Implement dql `Avg`\`Min`\`Max`

* feat: Implement `GroupBy` clause on dql

* fix(rbo): losing `InputRef` when pushing down `Project`

* refactor: primary key id is actively specified by the user

* fix: 'Alias' fails under aggregate function

* feat: Support `DateTime` type

+++ src/types/tuple.rs
@@ -7,7 +7,7 @@ use crate::types::value::{DataValue, ValueRef};
 
 const BITS_MAX_INDEX: usize = 8;
 
-pub type TupleId = usize;
+pub type TupleId = i64;
 
 #[derive(Clone, Debug, PartialEq)]
 pub struct Tuple {
@@ -51,7 +51,7 @@ impl Tuple {
         }
     }
 
-    /// e.g.: bits(u8)..|data_0(len for utf8_1)|utf8_0|data_1|f64_1|
+    /// e.g.: bits(u8)..|data_0(len for utf8_1)|utf8_0|data_1|
     /// Tips: all len is u32
     pub fn serialize_to(&self) -> Vec<u8> {
         fn flip_bit(bits: u8, i: usize) -> u8 {
